+++ src/Builder/DE/Concrete/FacturaElectronicaBuilder.php
@@ -16,6 +16,7 @@ use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GEmisTag;
 use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GOpeComTag;
 use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GOpeDETag;
 use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GPaConEIniTag;
+use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GPagTarCDTag;
 use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GRespDETag;
 use Nyxcode\PhpSifenTool\Composite\Concrete\DE\GTimbTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\CActEcoTag;
@@ -32,6 +33,7 @@ use Nyxcode\PhpSifenTool\Composite\Leaf\DE\CTipRegTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DAnoContTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCarRespDETag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCelRecTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCodAuOpeTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCodClienteTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCodSegTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DCompDir1Tag;
@@ -43,6 +45,7 @@ use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesActEcoTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesCiuEmiTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesCiuRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesCondAntTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesDenTarjTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesDepEmiTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesDepRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDesDisEmiTag;
@@ -62,6 +65,7 @@ use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDTipIDRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDTipIDRespDETag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDVEmiTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDVIdTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDVProTarTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DDVRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DEmailETag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DEmailRecTag;
@@ -80,14 +84,18 @@ use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNomFanEmiTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNomFanRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNomRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNomRespDETag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNomTitTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumCasRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumCasTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumDocTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumIDRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumIDRespDETag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumTarjTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DNumTimTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DPunExpTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DRSProTarTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DRucEmTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DRUCProTarTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DRucRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DSecContTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DSerieNumTag;
@@ -98,6 +106,8 @@ use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DTiCamTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\DTiCamTiPagTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\ICondAntTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\ICondOpeTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\IDenTarjTag;
+use Nyxcode\PhpSifenTool\Composite\Leaf\DE\IForProPaTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\IIndPresTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\INatRecTag;
 use Nyxcode\PhpSifenTool\Composite\Leaf\DE\ITiContRecTag;
@@ -116,6 +126,7 @@ use Nyxcode\PhpSifenTool\Enums\TipoCambioOperacion;
 use Nyxcode\PhpSifenTool\Enums\TipoCondicionOperacion;
 use Nyxcode\PhpSifenTool\Enums\TipoDocumentoElectronico;
 use Nyxcode\PhpSifenTool\Enums\TipoOperacion;
+use Nyxcode\PhpSifenTool\Enums\TipoPago;
 
 class FacturaElectronicaBuilder implements BuilderInterface
 {
@@ -129,6 +140,7 @@ class FacturaElectronicaBuilder implements BuilderInterface
     protected GDTipDETag $gDTipDE;
     protected GCamFETag $gCamFE;
     protected GCamCondTag $gCamCond;
+    protected GPaConEIniTag $gPaConEIni;
 
     public function reset()
     {
@@ -623,31 +635,86 @@ class FacturaElectronicaBuilder implements BuilderInterface
 
     public function setGroupE71($data)
     {
-        $gPaConEIni = new GPaConEIniTag();
+        $this->gPaConEIni = new GPaConEIniTag();
 
         $iTiPago = new ITiPagoTag($data["iTiPago"]);
-        $gPaConEIni->add($iTiPago);
+        $this->gPaConEIni->add($iTiPago);
 
         $dDesTiPag = new DDesTiPagTag($data["dDesTiPag"]);
-        $gPaConEIni->add($dDesTiPag);
+        $this->gPaConEIni->add($dDesTiPag);
 
         $dMonTiPag = new DMonTiPagTag($data["dMonTiPag"]);
-        $gPaConEIni->add($dMonTiPag);
+        $this->gPaConEIni->add($dMonTiPag);
 
         $cMoneTiPag = new CMoneTiPagTag($data["cMoneTiPag"]);
-        $gPaConEIni->add($cMoneTiPag);
+        $this->gPaConEIni->add($cMoneTiPag);
 
         $dDMoneTiPag = new DDMoneTiPagTag($data["dDMoneTiPag"]);
-        $gPaConEIni->add($dDMoneTiPag);
+        $this->gPaConEIni->add($dDMoneTiPag);
 
         if (
             $data["dTiCamTiPag"] &&
             $data["cMoneTiPag"] != MonedaOperacion::PYG->value
         ) {
             $dTiCamTiPag = new DTiCamTiPagTag($data["dTiCamTiPag"]);
-            $gPaConEIni->add($dTiCamTiPag);
+            $this->gPaConEIni->add($dTiCamTiPag);
         }
-        $this->gCamCond->add($gPaConEIni);
+
+        if (
+            $data["iTiPago"] == TipoPago::TARJETA_CREDITO->value ||
+            $data["iTiPago"] == TipoPago::TARJETA_DEBITO->value
+        ) {
+            $this->setGroupE711($data);
+        }
+
+        $this->gCamCond->add($this->gPaConEIni);
+    }
+
+    public function setGroupE711($data)
+    {
+        $gPagTarCD = new GPagTarCDTag();
+
+        $iDenTarj = new IDenTarjTag($data["iDenTarj"]);
+        $gPagTarCD->add($iDenTarj);
+
+        $dDesDenTarj = new DDesDenTarjTag($data["dDesDenTarj"]);
+        $gPagTarCD->add($dDesDenTarj);
+
+        if ($data["dRSProTar"]) {
+            $dRSProTar = new DRSProTarTag($data["dRSProTar"]);
+            $gPagTarCD->add($dRSProTar);
+        }
+
+        if (
+            isset($data["dRUCProTar"]) &&
+            isset($data["dDVProTar"])
+        ) {
+            $dRUCProTar = new DRUCProTarTag($data["dRUCProTar"]);
+            $gPagTarCD->add($dRUCProTar);
+
+            $dDVProTar = new DDVProTarTag($data["dDVProTar"]);
+            $gPagTarCD->add($dDVProTar);
+        }
+
+        $iForProPa = new IForProPaTag($data["iForProPa"]);
+        $gPagTarCD->add($iForProPa);
+
+        if ($data["dCodAuOpe"]) {
+            $dCodAuOpe = new DCodAuOpeTag($data["dCodAuOpe"]);
+            $gPagTarCD->add($dCodAuOpe);
+        }
+
+        if ($data["dNomTit"]) {
+            $dNomTit = new DNomTitTag($data["dNomTit"]);
+            $gPagTarCD->add($dNomTit);
+        }
+
+        if ($data["dNumTarj"]) {
+            $dNumTarj = new DNumTarjTag($data["dNumTarj"]);
+            $gPagTarCD->add($dNumTarj);
+        }
+
+        $this->gPaConEIni->add($gPagTarCD);
     }
 
     public function getResult()